requirements.txt
slackclient
psutil
subprocess.run
requests